work-breakdown/requirements.txt
openai==1.55.3
python-dotenv==1.0.0
langchain==0.1.16
langchain-openai==0.1.3
pydantic==1.10.8
PyPDF2==3.0.1
langchain-text-splitters==0.0.1